[home] Add auth to APIv2 client & use it for dev sessions
Adds auth support to the APIv2 client so that it sends the session secret in the `expo-session` header (if the session secret is defined). Use the API v2 client to fetch in-development project sessions instead of doing a raw fetch so that we can use all the APIv2 parsing code.

Tested by verifying that I could see my in-development projects based on my user account i.e. auth was working.